control_volume.sh -> control_volume.
I started cleaning up my machines. Having a file name extension like
this is useless and harmful. Without an extension, I can easily rewrite
that script in Python, C, Java or whatever. But *with* an extension, I'm
condemned to use a shell script (because I don't want to have a file
called "foo.sh" on my hard drive which is not a shell script).